Replace find_module with find_spec in util.imports.ModuleFinder
All that has changed from find_module to find_spec is the return type (must be importlib.machinery.ModuleSpec instead of a Loader).
This seems to mostly fix the issues with this module, however load_module is still used which is scheduled to be removed in python 3.15. I still haven't figured out how to update that part.