[adapters._django_db_models_base] Remove most of the 0.x and 1.x compatibility and fix a couple of problems
Most people won't be using 0.x and 1.x because it is too old (Django is on version 5.x now).
Fix: In DjangoClassAlias, use try/except when trying to delete a key because it is sometimes already removed
Fix: Look for cached things with object._state.fields_cache.{name} instead of object.__dict__._{name}_cache. I'm not sure which version this changed in.